refactor: rename @navita/swc → @navita/extraction
The package no longer uses swc (it's pure-JS on oxc-parser/oxc-transform), so the
name was misleading. Rename the package + dir, update @navita/core's dependency
and its single import, the vitest project map, and the changeset. CHANGELOG
history and the benchmark's reference to the old published @navita/swc@0.1.0 are
left as-is. (@navita/swc to be `npm deprecate`d after the renamed package
publishes.) Build/test/type-check all green.
