+++ packages/core/package.json
@@ -35,13 +35,13 @@
     "type-check": "tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
-    "@navita/types": "workspace:*",
     "@navita/adapter": "workspace:*",
-    "@navita/swc": "workspace:*",
     "@navita/engine": "workspace:*",
+    "@navita/extraction": "workspace:*",
+    "@navita/types": "workspace:*",
     "enhanced-resolve": "5.15.0",
-    "outdent": "^0.8.0",
-    "magic-string": "^0.30.3"
+    "magic-string": "^0.30.3",
+    "outdent": "^0.8.0"
   },
   "devDependencies": {
     "@navita/css": "workspace:*"
