+++ packages/extraction/src/stripTypes.ts
