+++ packages/extraction/src/toAmd.ts
